LayoutTests/platform/win/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/color-expected.txt


FAIL Empty value should return #000000 assert_equals: expected "#000000" but got ""
FAIL Missing value should return #000000 assert_equals: expected "#000000" but got "undefined"
PASS Valid simple color: should return #ffffff
FAIL Valid simple color (containing LATIN CAPITAL LETTERS): should return #ffffff (converted to ASCII lowercase) assert_equals: expected "#ffffff" but got "#FFFFFF"
FAIL Zero-padding assert_equals: expected "#0f0f0f" but got "#0F0F0F"
FAIL Invalid simple color: not 7 characters long assert_equals: expected "#000000" but got "#fff"
FAIL Invalid simple color: no starting # sign assert_equals: expected "#000000" but got "fffffff"
FAIL Invalid simple color: non ASCII hex digits assert_equals: expected "#000000" but got "#gggggg"
FAIL Invalid simple color: foobar assert_equals: expected "#000000" but got "foobar"
FAIL Invalid color: trailing Null (U+0000) assert_equals: expected "#000000" but got "#ffffff"
FAIL Invalid color: trailing ; assert_equals: expected "#000000" but got "#ffffff;"
FAIL Invalid color: leading space assert_equals: expected "#000000" but got " #ffffff"
FAIL Invalid color: trailing space assert_equals: expected "#000000" but got "#ffffff "
FAIL Invalid color: leading+trailing spaces assert_equals: expected "#000000" but got " #ffffff "
FAIL Invalid color: keyword crimson assert_equals: expected "#000000" but got "crimson"
FAIL Invalid color: keyword bisque assert_equals: expected "#000000" but got "bisque"
FAIL Invalid color: keyword currentColor assert_equals: expected "#000000" but got "currentColor"
FAIL Invalid color: keyword transparent assert_equals: expected "#000000" but got "transparent"
FAIL Invalid color: keyword ActiveBorder assert_equals: expected "#000000" but got "ActiveBorder"
FAIL Invalid color: keyword inherit assert_equals: expected "#000000" but got "inherit"
FAIL Invalid color: rgb(1,1,1) assert_equals: expected "#000000" but got "rgb(1,1,1)"
FAIL Invalid color: rgb(1,1,1,1) assert_equals: expected "#000000" but got "rgb(1,1,1,1)"
FAIL Invalid color: PILE OF POO (U+1F4A9) assert_equals: expected "#000000" but got "#FFFFFὊ9"
